slightly rewrite the tests
