+++ development/tools/jobs/incubator/third-party-images/cloudsql_proxy_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-const cloudSQLProxyJobPath = "./../../../../../../prow/jobs/incubator/third-party-images/cloudsql-proxy/cloudsql-proxy.yaml"
+const cloudSQLProxyJobPath = "../../../../../prow/jobs/incubator/third-party-images/third-party-images.yaml"
 
 func TestCloudSQLProxyJobsPresubmit(t *testing.T) {
 	// when
@@ -25,11 +25,10 @@ func TestCloudSQLProxyJobsPresubmit(t *testing.T) {
 	assert.True(t, actualPresubmit.Decorate)
 	assert.False(t, actualPresubmit.Optional)
 	assert.Equal(t, "^cloudsql-proxy/", actualPresubmit.RunIfChanged)
-	assert.Equal(t, "github.com/kyma-incubator/third-party-images", actualPresubmit.PathAlias)
-	tester.AssertThatHasPresets(t, actualPresubmit.JobBase, preset.DindEnabled, preset.DockerPushRepoIncubator, preset.GcrPush, preset.BuildPr)
-	assert.Equal(t, tester.ImageBootstrap20181204, actualPresubmit.Spec.Containers[0].Image)
-	assert.Equal(t, []string{"/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/build.sh"}, actualPresubmit.Spec.Containers[0].Command)
-	assert.Equal(t, []string{"/home/prow/go/src/github.com/kyma-incubator/third-party-images/cloudsql-proxy"}, actualPresubmit.Spec.Containers[0].Args)
+	tester.AssertThatHasPresets(t, actualPresubmit.JobBase, preset.DindEnabled, preset.DockerPushRepoIncubator, preset.GcrPush)
+	assert.Equal(t, tester.ImageBootstrapTestInfraLatest, actualPresubmit.Spec.Containers[0].Image)
+	assert.Equal(t, []string{"/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/build-generic.sh"}, actualPresubmit.Spec.Containers[0].Command)
+	assert.Equal(t, []string{"/home/prow/go/src/github.com/kyma-incubator/third-party-images/cloudsql-proxy", "ci-pr"}, actualPresubmit.Spec.Containers[0].Args)
 }
 
 func TestCloudSQLProxyJobPostsubmit(t *testing.T) {
@@ -44,9 +43,8 @@ func TestCloudSQLProxyJobPostsubmit(t *testing.T) {
 
 	assert.Equal(t, 10, actualPostsubmit.MaxConcurrency)
 	assert.True(t, actualPostsubmit.Decorate)
-	assert.Equal(t, "github.com/kyma-incubator/third-party-images", actualPostsubmit.PathAlias)
-	tester.AssertThatHasPresets(t, actualPostsubmit.JobBase, preset.DindEnabled, preset.DockerPushRepoIncubator, preset.GcrPush, preset.BuildMaster)
-	assert.Equal(t, tester.ImageBootstrap20181204, actualPostsubmit.Spec.Containers[0].Image)
-	assert.Equal(t, []string{"/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/build.sh"}, actualPostsubmit.Spec.Containers[0].Command)
-	assert.Equal(t, []string{"/home/prow/go/src/github.com/kyma-incubator/third-party-images/cloudsql-proxy"}, actualPostsubmit.Spec.Containers[0].Args)
+	tester.AssertThatHasPresets(t, actualPostsubmit.JobBase, preset.DindEnabled, preset.DockerPushRepoIncubator, preset.GcrPush)
+	assert.Equal(t, tester.ImageBootstrapTestInfraLatest, actualPostsubmit.Spec.Containers[0].Image)
+	assert.Equal(t, []string{"/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/build-generic.sh"}, actualPostsubmit.Spec.Containers[0].Command)
+	assert.Equal(t, []string{"/home/prow/go/src/github.com/kyma-incubator/third-party-images/cloudsql-proxy", "ci-master"}, actualPostsubmit.Spec.Containers[0].Args)
 }
