+++ development/tools/jobs/incubator/third-party-images/kiwigrid_sidecar_test.go
@@ -1,0 +1,49 @@
+package third_party_images_test
+
+import (
+	"testing"
+
+	"github.com/kyma-project/test-infra/development/tools/jobs/tester"
+	"github.com/kyma-project/test-infra/development/tools/jobs/tester/preset"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+const kiwigridSidecarJobPath = "../../../../../prow/jobs/incubator/third-party-images/third-party-images.yaml"
+
+func TestKiwigridSidecarJobsPresubmit(t *testing.T) {
+	// when
+	jobConfig, err := tester.ReadJobConfig(kiwigridSidecarJobPath)
+
+	// then
+	require.NoError(t, err)
+	actualPresubmit := tester.FindPresubmitJobByNameAndBranch(jobConfig.AllStaticPresubmits([]string{"kyma-incubator/third-party-images"}), "pre-master-tpi-kiwigrid-sidecar", "master")
+	require.NotNil(t, actualPresubmit)
+
+	assert.Equal(t, 10, actualPresubmit.MaxConcurrency)
+	assert.False(t, actualPresubmit.SkipReport)
+	assert.True(t, actualPresubmit.Decorate)
+	assert.Equal(t, "^kiwigrid-sidecar/", actualPresubmit.RunIfChanged)
+	tester.AssertThatHasPresets(t, actualPresubmit.JobBase, preset.DindEnabled, preset.DockerPushRepoIncubator, preset.GcrPush)
+	assert.Equal(t, tester.ImageBootstrapTestInfraLatest, actualPresubmit.Spec.Containers[0].Image)
+	assert.Equal(t, []string{"/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/build-generic.sh"}, actualPresubmit.Spec.Containers[0].Command)
+	assert.Equal(t, []string{"/home/prow/go/src/github.com/kyma-incubator/third-party-images/kiwigrid-sidecar", "ci-pr"}, actualPresubmit.Spec.Containers[0].Args)
+}
+
+func TestKiwigridSidecarJobPostsubmit(t *testing.T) {
+	// when
+	jobConfig, err := tester.ReadJobConfig(kiwigridSidecarJobPath)
+
+	// then
+	require.NoError(t, err)
+
+	actualPostsubmit := tester.FindPostsubmitJobByNameAndBranch(jobConfig.AllStaticPostsubmits([]string{"kyma-incubator/third-party-images"}), "post-master-tpi-kiwigrid-sidecar", "master")
+	require.NotNil(t, actualPostsubmit)
+
+	assert.Equal(t, 10, actualPostsubmit.MaxConcurrency)
+	assert.True(t, actualPostsubmit.Decorate)
+	tester.AssertThatHasPresets(t, actualPostsubmit.JobBase, preset.DindEnabled, preset.DockerPushRepoIncubator, preset.GcrPush)
+	assert.Equal(t, tester.ImageBootstrapTestInfraLatest, actualPostsubmit.Spec.Containers[0].Image)
+	assert.Equal(t, []string{"/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/build-generic.sh"}, actualPostsubmit.Spec.Containers[0].Command)
+	assert.Equal(t, []string{"/home/prow/go/src/github.com/kyma-incubator/third-party-images/kiwigrid-sidecar", "ci-master"}, actualPostsubmit.Spec.Containers[0].Args)
+}
